converter/pyproject.toml
[project]
name = "converter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.0"
dependencies = [
    "aioboto3>=15.1.0",
    "aiokafka>=0.12.0",
    "certifi>=2025.8.3",
    "dishka>=1.8.0",
    "faststream[cli]==0.5.39",
    "ffmpeg-python==0.2.0",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pydantic-settings>=2.13.1",
    "prometheus-client>=0.21.0",
    "python-dotenv==1.0.1",
    "types-aiobotocore-s3>=3.1.3",
]

[dependency-groups]
dev = [
    "ruff>=0.15.2",
    "ty>=0.0.18",
]

[tool.ruff]
line-length = 120

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    ".venv",
    "**/migrations/**",
]

[tool.ruff.lint]
select = [
    "ALL"
]
ignore = [
    "D103",
    "D102",
    "D100",
    "D101",
    "D104",
    "D107",
    "D105",
    "ANN401",
    "ANN204",
    "ANN202",
    "ANN201",
    "ANN003",
    "ANN002",
    "N815",
    "TC001",
    "D203",
    "D212",
    "PLR2004",
    "TRY301",
    "RUF006",
    "PGH003",
    "RUF002",
    "D415",
]


[tool.ruff.lint.isort]
known-first-party = [
    "application",
    "controllers",
    "infrastructure",
    "config",
    "ioc",
    "main",
]

[tool.ty.environment]
python-version = "3.12"
root = ["src"]

[tool.ty.src]
exclude = ["**/migrations/**"]